Move unit tests to subfolders
