+++ src/state/locales/test/reducer.js
