+++ src/state/middlewares/socket-io/test/agents.js
@@ -1,10 +1,10 @@
-import { ok, equal, deepEqual } from 'assert'
-import mockIO from '../mock-io'
-import agentMiddleware from 'state/middlewares/socket-io/agents'
-import { AGENT_INBOUND_MESSAGE } from 'state/action-types'
-import { agentReceiveMessage } from 'state/chatlist/actions'
+import { ok, equal, deepEqual } from 'assert';
+import agentMiddleware from 'state/middlewares/socket-io/agents';
+import { AGENT_INBOUND_MESSAGE } from 'state/action-types';
+import { agentReceiveMessage } from 'state/chatlist/actions';
+import mockIO from '../../../../../test/mocks/mock-io';
 
-const debug = require( 'debug' )( 'happychat:test:agent' )
+const debug = require( 'debug' )( 'happychat:test:agent' );
 
 describe( 'Agent Service', () => {
 	let server, socket, client, io
