+++ src/state/middlewares/system/test/index.js
@@ -1,0 +1,3 @@
+describe( __filename.replace( global.BASE_PATH, '' ), () => {
+	it( 'should really need some tests', () => {} );
+} );
