--- test/tick.js
@@ -1,1 +1,0 @@
-export const tick = ( fn ) => ( ... args ) => process.nextTick( () => fn( ... args ) )
